app/requirements.txt
openai
sentence_transformers 
streamlit
requests
groq 
pinecone
psycopg2-binary==2.9.9
python-dotenv